requirements.txt
gcsfs
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-bigquery[pandas]
google-cloud-resource-manager
google-cloud-storage
huggingface_hub
pandas
pydicom~=2.4.4
python-dotenv
numpy~=1.26.4
matplotlib
seaborn
scikit-learn~=1.5.1
timm
tqdm
torch
transformers
pillow~=10.4.0
openpyxl